requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic-settings
pydantic[email]
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
scikit-learn
pandas
python-multipart
joblib